app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==4.2.15
django-crontab==0.7.1
django-environ==0.11.2
django-guardian==2.4.0
django-oauth-toolkit==2.4.0
django-rest-knox==5.0.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
